+++ src/Field.php
@@ -51,6 +51,10 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
     const PROPAGATION_METHOD_NONE = 'none';
     const PROPAGATION_METHOD_SITE_GROUP = 'siteGroup';
     const PROPAGATION_METHOD_LANGUAGE = 'language';
+    /**
+     * @since 2.12.0
+     */
+    const PROPAGATION_METHOD_CUSTOM = 'custom';
     const PROPAGATION_METHOD_ALL = 'all';
 
     /**
@@ -124,10 +128,11 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
      *
      * This will be set to one of the following:
      *
-     * - `none` – Only save b locks in the site they were created in
-     * - `siteGroup` – Save  blocks to other sites in the same site group
+     * - `none` – Only save blocks in the site they were created in
+     * - `siteGroup` – Save blocks to other sites in the same site group
      * - `language` – Save blocks to other sites with the same language
      * - `all` – Save blocks to all sites supported by the owner element
+     * - `custom` – Save blocks to sites depending on the [[propagationKeyFormat]] value
      *
      * @since 2.4.0
      */
@@ -137,6 +142,12 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
      * @var string The old propagation method for this field
      */
     private $_oldPropagationMethod;
+
+    /**
+     * @var string|null The field’s propagation key format, if [[propagationMethod]] is `custom`
+     * @since 2.12.0
+     */
+    public $propagationKeyFormat;
 
     /**
      * @inheritdoc
@@ -155,6 +166,10 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
      */
     public function init()
     {
+        if ($this->propagationKeyFormat === '') {
+            $this->propagationKeyFormat = null;
+        }
+
         // Set localizeBlocks in case anything is still checking it
         $this->localizeBlocks = $this->propagationMethod === self::PROPAGATION_METHOD_NONE;
         parent::init();
@@ -181,6 +196,7 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
                 self::PROPAGATION_METHOD_NONE,
                 self::PROPAGATION_METHOD_SITE_GROUP,
                 self::PROPAGATION_METHOD_LANGUAGE,
+                self::PROPAGATION_METHOD_CUSTOM,
                 self::PROPAGATION_METHOD_ALL
             ]
         ];
@@ -660,8 +676,9 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
 
         if ($this->oldSettings !== null) {
             $oldPropagationMethod = $this->oldSettings['propagationMethod'] ?? self::PROPAGATION_METHOD_ALL;
+            $oldPropagationKeyFormat = $this->oldSettings['propagationKeyFormat'] ?? null;
 
-            if ($this->propagationMethod !== $oldPropagationMethod) {
+            if ($this->propagationMethod !== $oldPropagationMethod || $this->propagationKeyFormat !== $oldPropagationKeyFormat) {
                 Queue::push(new ApplyNewPropagationMethod([
                     'description' => Craft::t('neo', 'Applying new propagation method to Neo blocks'),
                     'elementType' => Block::class,