+++ src/services/Fields.php
@@ -273,7 +273,7 @@ class Fields extends Component
                 ($owner->propagateAll || !empty($owner->newSiteIds))
             ) {
                 $ownerSiteIds = ArrayHelper::getColumn(ElementHelper::supportedSitesForElement($owner), 'siteId');
-                $fieldSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $owner);
+                $fieldSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $owner, $field->propagationKeyFormat);
                 $otherSiteIds = array_diff($ownerSiteIds, $fieldSiteIds);
 
                 if (!$owner->propagateAll) {
@@ -307,7 +307,7 @@ class Fields extends Component
                         }
                         $this->duplicateBlocks($field, $owner, $otherTarget);
                         // Make sure we don't duplicate blocks for any of the sites that were just propagated to
-                        $sourceSupportedSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $otherTarget);
+                        $sourceSupportedSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $otherTarget, $field->propagationKeyFormat);
                         $handledSiteIds = array_merge($handledSiteIds, array_flip($sourceSupportedSiteIds));
                     }
                     $owner->setFieldValue($field->handle, $query);
@@ -439,7 +439,7 @@ class Fields extends Component
         if ($checkOtherSites && $field->propagationMethod !== Field::PROPAGATION_METHOD_ALL) {
             // Find the target's site IDs that *aren't* supported by this site's neo blocks
             $targetSiteIds = ArrayHelper::getColumn(ElementHelper::supportedSitesForElement($target), 'siteId');
-            $fieldSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $target);
+            $fieldSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $target, $field->propagationKeyFormat);
             $otherSiteIds = array_diff($targetSiteIds, $fieldSiteIds);
 
             if (!empty($otherSiteIds)) {
@@ -481,7 +481,7 @@ class Fields extends Component
                     $this->duplicateBlocks($field, $otherSource, $otherTargets[$otherSource->siteId]);
 
                     // Make sure we don't duplicate blocks for any of the sites that were just propagated to
-                    $sourceSupportedSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $otherSource);
+                    $sourceSupportedSiteIds = $this->getSupportedSiteIds($field->propagationMethod, $otherSource, $field->propagationKeyFormat);
                     $handledSiteIds = array_merge($handledSiteIds, $sourceSupportedSiteIds);
                 }
             }
@@ -633,7 +633,7 @@ class Fields extends Component
                 $this->_saveNeoStructuresForSites($field, $owner, $newBlocks, $canonicalOwner->siteId);
             }
 
-            $siteIds = $this->getSupportedSiteIds($field->propagationMethod, $canonicalOwner);
+            $siteIds = $this->getSupportedSiteIds($field->propagationMethod, $canonicalOwner, $field->propagationKeyFormat);
 
             foreach ($siteIds as $siteId) {
                 $handledSiteIds[$siteId] = true;
@@ -652,7 +652,7 @@ class Fields extends Component
      */
     public function getSupportedSiteIdsForField(Field $field, ElementInterface $owner): array
     {
-        return $this->getSupportedSiteIds($field->propagationMethod, $owner);
+        return $this->getSupportedSiteIds($field->propagationMethod, $owner, $field->propagationKeyFormat);
     }
 
     /**
@@ -660,17 +660,24 @@ class Fields extends Component
      *
      * @param string $propagationMethod
      * @param ElementInterface $owner
+     * @param string|null $propagationKeyFormat
      * @return int[]
      * @throws
      * @since 2.5.10
      */
-    public function getSupportedSiteIds(string $propagationMethod, ElementInterface $owner): array
+    public function getSupportedSiteIds(string $propagationMethod, ElementInterface $owner, ?string $propagationKeyFormat = null): array
     {
         /** @var Element $owner */
         /** @var Site[] $allSites */
         $allSites = ArrayHelper::index(Craft::$app->getSites()->getAllSites(), 'id');
         $ownerSiteIds = ArrayHelper::getColumn(ElementHelper::supportedSitesForElement($owner), 'siteId');
         $siteIds = [];
+
+        if ($propagationMethod === Field::PROPAGATION_METHOD_CUSTOM && $propagationKeyFormat !== null) {
+            $view = Craft::$app->getView();
+            $elementsService = Craft::$app->getElements();
+            $propagationKey = $view->renderObjectTemplate($propagationKeyFormat, $owner);
+        }
 
         foreach ($ownerSiteIds as $siteId) {
             switch ($propagationMethod) {
@@ -682,6 +689,14 @@ class Fields extends Component
                     break;
                 case Field::PROPAGATION_METHOD_LANGUAGE:
                     $include = $allSites[$siteId]->language == $allSites[$owner->siteId]->language;
+                    break;
+                case Field::PROPAGATION_METHOD_CUSTOM:
+                    if (!isset($propagationKey)) {
+                        $include = true;
+                    } else {
+                        $siteOwner = $elementsService->getElementById($owner->id, get_class($owner), $siteId);
+                        $include = $siteOwner && $propagationKey === $view->renderObjectTemplate($propagationKeyFormat, $siteOwner);
+                    }
                     break;
                 default:
                     $include = true;
@@ -700,7 +715,7 @@ class Fields extends Component
     {
         // we need to setup the structure for the other supported sites too.
         // must be immediate to show changes on the front end.
-        $supported = $this->getSupportedSiteIds($field->propagationMethod, $owner);
+        $supported = $this->getSupportedSiteIds($field->propagationMethod, $owner, $field->propagationKeyFormat);
 
         // remove the current
         if (($key = array_search($owner->siteId, $supported)) !== false) {
@@ -732,7 +747,7 @@ class Fields extends Component
      */
     private function _deleteOtherBlocks(Field $field, ElementInterface $owner, array $except)
     {
-        $supportedSites = $this->getSupportedSiteIds($field->propagationMethod, $owner);
+        $supportedSites = $this->getSupportedSiteIds($field->propagationMethod, $owner, $field->propagationKeyFormat);
         $supportedSitesCount = count($supportedSites);
 
         if ($supportedSitesCount > 1 && $field->propagationMethod !== Field::PROPAGATION_METHOD_NONE) {
@@ -807,7 +822,7 @@ class Fields extends Component
         Neo::$plugin->blocks->buildStructure($blocks, $blockStructure);
 
         // if multi site then save the structure for it. since it's all the same then we can use the same structure.
-        $supported = $this->getSupportedSiteIdsExCurrent($field, $owner);
+        $supported = $this->getSupportedSiteIdsExCurrent($field, $owner, $field->propagationKeyFormat);
         $supportedCount = count($supported);
 
         if ($supportedCount > 0) {